Implement Minters, Splitters and Burners

+++ src/TokenSystem/Exceptions/InsufficientTokenAmountException.cs
@@ -1,12 +1,13 @@
 // Copyright (c) Comrade Coop. All rights reserved.
 
 using System;
+using System.Numerics;
 
 namespace TokenSystem.Exceptions
 {
 	public class InsufficientTokenAmountException : Exception
 	{
-		public InsufficientTokenAmountException(decimal total, decimal requested)
+		public InsufficientTokenAmountException(BigInteger total, BigInteger requested)
 			: base($"Insufficient token amount requested: Requested {requested} from only {total}.")
 		{
 		}
