+++ src/TokenSystem/TokenEventArgs/TokensTransferredEventArgs.cs
@@ -1,6 +1,7 @@
 // Copyright (c) Comrade Coop. All rights reserved.
 
 using System;
+using System.Numerics;
 using ContractsCore;
 using TokenSystem.Tokens;
 
@@ -9,7 +10,7 @@ namespace TokenSystem.TokenEventArgs
 	public class TokensTransferredEventArgs<TTagType> : EventArgs
 	{
 		public TokensTransferredEventArgs(
-			decimal amount,
+			BigInteger amount,
 			IReadOnlyTaggedTokens<TTagType> tokens,
 			Address from,
 			Address to)
@@ -20,7 +21,7 @@ namespace TokenSystem.TokenEventArgs
 			this.To = to;
 		}
 
-		public decimal Amount { get; }
+		public BigInteger Amount { get; }
 
 		public IReadOnlyTaggedTokens<TTagType> Tokens { get; }
