+++ src/TokenSystem/TokenManager/FungibleTokenTagger.cs
@@ -1,6 +1,7 @@
 // Copyright (c) Comrade Coop. All rights reserved.
 
 using System.Collections.Generic;
+using System.Numerics;
 using ContractsCore;
 using TokenSystem.Exceptions;
 using TokenSystem.Tokens;
@@ -11,14 +12,14 @@ namespace TokenSystem.TokenManager
 	{
 		public const string DefaultTokenTag = "FungibleToken";
 
-		public IReadOnlyTaggedTokens<string> Tag(Address owner, decimal amount)
+		public IReadOnlyTaggedTokens<string> Tag(Address owner, BigInteger amount)
 		{
 			if (amount < 0)
 			{
 				throw new NonPositiveTokenAmountException(amount);
 			}
 
-			var tokens = new ReadOnlyTaggedTokens<string>(new SortedDictionary<string, decimal>
+			var tokens = new ReadOnlyTaggedTokens<string>(new SortedDictionary<string, BigInteger>
 			{
 				[DefaultTokenTag] = amount,
 			});
