+++ src/TokenSystem/TokenManager/ITokenTagger.cs
@@ -1,5 +1,6 @@
 // Copyright (c) Comrade Coop. All rights reserved.
 
+using System.Numerics;
 using ContractsCore;
 using TokenSystem.Tokens;
 
@@ -7,6 +8,6 @@ namespace TokenSystem.TokenManager
 {
 	public interface ITokenTagger<TTagType>
 	{
-		IReadOnlyTaggedTokens<TTagType> Tag(Address owner, decimal amount);
+		IReadOnlyTaggedTokens<TTagType> Tag(Address owner, BigInteger amount);
 	}
 }
